perf(npm): only install cypress when needed
We run `npm ci` in a number of places
and usually we do not need cypress to be installed.

In text a sample run of `npm ci` without cypress took 26 sec. instead of 46 sec.

This also prevents run failures due to network issues connecting to cypress.

Documentation for the environment variable used:
https://docs.cypress.io/guides/references/advanced-installation#Skipping-installation

Signed-off-by: Max <max@nextcloud.com>