fix(CI): Make sure the "when unrelated" fake summaries work correctly
Before they executed when any file outside the defined scope matched.
However that means: When a PR touches PHP and JS, both the real and
the fake summaries for node, eslint and phpunit got triggered.
Since the fake summary is always green it meant one could merge a PR
that in the end had red CI on the non-fake summary.

Now the fake summaries are skipped, if any of the files still matches
the pattern, instead of being executed whenever a single file did not
match it.

Signed-off-by: Joas Schilling <coding@schilljs.com>