fix(phpunit-mariadb): Use correct script name and env variable
In MariaDB 11 there was a "breaking" change in the docker files that
removed the legacy `mysqladm` and replaced it with `mariadb-admin`.
The symlink is not provided in the docker image so we need to use the
correct name here to make 10.6 AND 11.4 work.

Also the `MYSQL_` env variables are deprecated and will be removed.

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>